update feed like and socket event

+++ Controllers/FeedController.cs
@@ -1,5 +1,4 @@
 using API.DTOs;
-using API.Entities;
 using API.Interfaces;
 using AutoMapper;
 using Microsoft.AspNetCore.Authorization;
@@ -12,12 +11,19 @@ namespace API.Controllers;
 public class FeedController: BaseApiController {
     private readonly IUserService _userService;
     private readonly IFeedService _feedService;
+    private readonly IFeedCommentService _feedCommentService;
     private readonly IMapper _mapper;
 
-    public FeedController(IUserService userService, IFeedService feedService, IMapper mapper) {
+    public FeedController(
+     IUserService userService, 
+     IFeedService feedService, 
+     IMapper mapper,
+     IFeedCommentService feedCommentService
+   ) {
         _userService = userService;
         _feedService = feedService;
         _mapper = mapper;
+        _feedCommentService = feedCommentService;
     }
 
     [HttpPost("")]
@@ -42,9 +48,46 @@ public class FeedController: BaseApiController {
          return feeds;
     }
 
+    [HttpPut("{id}")]
+    public async Task<ActionResult<BooleanReturnDto>> UpdateFeed (int id, UpdateFeedDto data) {
+         int userId = _userService.GetConnectedUser(User);
+
+         BooleanReturnDto? result = await _feedService.UpdateFeed(id, userId, data);
+
+         if(result is null) return BadRequest("An Error occured. Can't update the feed");
+         if(result.Status == false) return NotFound(result.Message);
+
+         return result;
+    }
+
+    [HttpPost("{id}/like")]
+    public async Task<ActionResult<BooleanReturnDto>> AddFeedLike (int id) {
+         int userId = _userService.GetConnectedUser(User);
+
+         BooleanReturnDto? feed = await _feedService.AddFeedLikes(id, userId);
+
+         if(feed is null) return BadRequest("Couldn't create the feed like. please retry later");
+
+         return feed;
+    }
+
+    [HttpDelete("{id}/like")]
+    public async Task<ActionResult<BooleanReturnDto>> DeleteFeedLike (int id) {
+         int userId = _userService.GetConnectedUser(User);
+
+         BooleanReturnDto? feed = await _feedService.RemoveFeedLikes(id, userId);
+
+         if(feed is null) return BadRequest("Couldn't remove the feed like. please retry later");
+
+         return feed;
+    }
+
     [HttpDelete("{id}")]
     public async Task<ActionResult<BooleanReturnDto>> DeleteFeed (int id) {
          int userId = _userService.GetConnectedUser(User);
+
+         bool validFeedId = await _feedService.IsValidFeedId(id);
+         if(!validFeedId) return NotFound("Invalid Feed Id");
 
          BooleanReturnDto? result = await _feedService.DeleteFeed(id, userId);
 
@@ -52,5 +95,23 @@ public class FeedController: BaseApiController {
          if(result.Status == false) return NotFound(result.Message);
 
          return result;
+    }
+
+    [HttpGet("{id}/comments")]
+    public async Task<ActionResult<ResultPaginate<ResultFeedCommentDto>>> GetFeedComments (int id, int skip=0, int limit=10) {
+
+         ResultPaginate<ResultFeedCommentDto> comments = await _feedCommentService.GetFeedComments(id, skip, limit);
+
+         return comments;
+    }
+
+    [HttpPost("{id}/comments")]
+    public async Task<ActionResult<ResultFeedCommentDto>> CreateComment (CreateCommentDto data, int id) {
+       int userId = _userService.GetConnectedUser(User);
+       ResultFeedCommentDto? comment = await _feedCommentService.CreateComment(data, id, userId);
+
+       if(comment is null) return BadRequest("An Error occured. Can't create the feed comment");
+
+       return comment;
     }
 }
